# first-example/api-docs.json
{
	"openapi": "3.0.1",
	"info": {
		"title": "sample application API",
		"description": "some description",
		"termsOfService": "http://swagger.io/terms/",
		"license": {
			"name": "Apache 2.0",
			"url": "http://springdoc.org"
		},
		"version": "0.0.2-SNAPSHOT"
	},
	"servers": [
		{
			"url": "http://localhost:8080",
			"description": "Generated server url"
		}
	],
	"paths": {
...
		"/abc/{abc}": {
			"get": {
				"tags": [
					"another-controller"
				],
				"operationId": "another_1",
				"x-damah": false,
				"parameters": [
					{
						"name": "abc",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"*/*": {
								"schema": {
									"type": "object",
									"additionalProperties": {
										"type": "string"
									}
								}
							}
						}
					}
				}
			},
			"post": {
				"tags": [
					"another-controller"
				],
				"operationId": "another_2",
				"x-damah": false,
				"parameters": [
					{
						"name": "abc",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Person"
							},
							"examples": {
								"user": {
									"summary": "person Example",
									"externalValue": "classpath:examples/1.json"
								},
								"user1": {
									"summary": "person1 Example",
									"externalValue": "classpath:examples/2.json"
								}
							}
						},
						"application/xml": {
							"schema": {
								"$ref": "#/components/schemas/Person"
							},
							"examples": {
								"user": {
									"summary": "person Example",
									"externalValue": "classpath:examples/1.xml"
								},
								"user1": {
									"summary": "person1 Example",
									"externalValue": "classpath:examples/2.xml"
								}
							}
						},
						"application/x-www-form-urlencoded": {
							"schema": {
								"$ref": "#/components/schemas/Person"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"*/*": {
								"schema": {
									"$ref": "#/components/schemas/Person"
								}
							}
						}
					}
				}
			}
		},
		"/def": {
			"post": {
				"tags": [
					"another-controller"
				],
				"operationId": "another_3",
				"parameters": [
					{
						"name": "abc",
						"in": "query",
						"required": true,
						"schema": {
							"maximum": 30,
							"minimum": 18,
							"type": "integer",
							"format": "int32"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Person"
							},
							"examples": {
								"user": {
									"summary": "person Example",
									"externalValue": "classpath:examples/1.json"
								},
								"user1": {
									"summary": "person1 Example",
									"externalValue": "classpath:examples/2.json"
								}
							}
						},
						"application/xml": {
							"schema": {
								"$ref": "#/components/schemas/Person"
							},
							"examples": {
								"user": {
									"summary": "person Example",
									"externalValue": "classpath:examples/1.xml"
								},
								"user1": {
									"summary": "person1 Example",
									"externalValue": "classpath:examples/2.xml"
								}
							}
						},
						"application/x-www-form-urlencoded": {
							"schema": {
								"$ref": "#/components/schemas/Person"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Person"
								}
							}
						}
					}
				}
			}
		}
	},
	"components": {
...
		"schemas": {
			"Person": {
				"required": [
					"lastName"
				],
				"type": "object",
				"properties": {
					"id": {
						"type": "integer",
						"format": "int64"
					},
					"firstName": {
						"maxLength": 2147483647,
						"minLength": 2,
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"email": {
						"pattern": ".+@.+\\..+",
						"type": "string"
					},
					"email1": {
						"type": "string",
						"x-Email": true
					},
					"age": {
						"maximum": 30,
						"minimum": 18,
						"type": "integer",
						"format": "int32"
					},
					"creditCardNumber": {
						"type": "string",
						"x-CreditCardNumber": true
					},
					"registrationDate": {
						"type": "string",
						"format": "date"
					},
					"pic": {
						"type": "string",
						"format": "byte"
					},
					"pics": {
						"type": "array",
						"items": {
							"type": "string",
							"format": "byte"
						}
					},
					"sampleCustomTypeData": {
						"$ref": "#/components/schemas/SampleCustomType"
					},
					"someTimeData": {
						"type": "string",
						"format": "date-time"
					},
					"anotherPerson": {
						"$ref": "#/components/schemas/Person"
					},
					"children": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/Person"
						}
					}
				},
				"xml": {
					"name": "person"
				}
			},
			"SampleCustomType": {
				"type": "object"
			}
		}
	}
}