# first-example/src/main/java/com/example/custom/SampleCustomType.java
package com.example.custom;

public class SampleCustomType {
	private String data;

	public SampleCustomType(String data) {
		super();
		this.data = data;
	}

	public SampleCustomType() {
		super();
		// TODO Auto-generated constructor stub
	}

	public String getData() {
		return data;
	}

	public void setData(String data) {
		this.data = data;
	}

}